call_record_details/requirements.txt
dj-database-url==0.5.0
Django==1.11.4
django-environ==0.4.5
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
whitenoise==3.3.1
